homework6/services/billing/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2
drf-yasg